refactor: make 'soliplex.agui' a package
Split the event stream parser off from the persistence mechanism.

Hoist 'EventStreamParser' to new 'soliplex.agui.parser' module.

Move persisence mechanis from 'soliplex.agui' to new 'soliplex.agui.thread'
module.
