+++ src/soliplex/installation.py
@@ -6,11 +6,11 @@ import fastapi
 import pydantic_ai
 
 from soliplex import agents
-from soliplex import agui
 from soliplex import config
 from soliplex import convos
 from soliplex import mcp_server
 from soliplex import secrets
+from soliplex.agui import thread as agui_thread
 
 
 @dataclasses.dataclass
@@ -123,7 +123,7 @@ async def lifespan(
     the_installation.resolve_secrets()
     the_installation.resolve_environment()
     the_convos = convos.Conversations()
-    the_threads = agui.Threads()
+    the_threads = agui_thread.Threads()
 
     context = {
         "the_installation": the_installation,
